+++ examples/features/jhttptempl/templates/home.html
@@ -25,7 +25,8 @@
         </div>
 
         <form action="/setlang" class="col-lg-4 mx-auto my-5" method="get">
-            <select class="form-select" name="lang" aria-label="{{.i18n.Tr "Language selection field"}}" onchange="this.form.submit()">
+            <select aria-label="{{.i18n.Tr " class="form-select" home.selection_fieldname="lang""}}"
+            onchange="this.form.submit()">
             <option selected>{{.i18n.Tr "home.choose_language"}}</option>
             <option value="">Best match</option>
             <option value="en">English</option>
