+++ locale.go
@@ -4,9 +4,30 @@ import (
 	"golang.org/x/text/language"
 
 	"github.com/vorlif/spreak/catalog"
-	"github.com/vorlif/spreak/internal/poplural"
 	"github.com/vorlif/spreak/localize"
+	"github.com/vorlif/spreak/pkg/poplural"
 )
+
+// Deprecated: Will be removed with v1.0. A Localizer should be used, as it offers the same functionalities.
+type Locale = Localizer
+
+// NewLocale creates a new locale for a language and the default domain of the bundle.
+// If a locale is not found, an error is returned.
+// Deprecated: Will be removed with v1.0. Use NewLocalizer instead.
+func NewLocale(bundle *Bundle, lang language.Tag) (*Locale, error) {
+	return NewLocaleWithDomain(bundle, lang, bundle.defaultDomain)
+}
+
+// NewLocaleWithDomain creates a new locale for a language and a default domain.
+// If no locale is found, an error is returned.
+// Deprecated: Will be removed with v1.0. Use NewLocalizerForDomain instead.
+func NewLocaleWithDomain(bundle *Bundle, lang language.Tag, domain string) (*Locale, error) {
+	l := NewLocalizerForDomain(bundle, domain, lang)
+	if !l.HasLocale() {
+		return nil, newMissingLanguageError(lang)
+	}
+	return l, nil
+}
 
 type locale struct {
 	bundle           *Bundle
