+++ pkg/README.md
@@ -1,0 +1,3 @@
+The following packages are subject to change and are not intended primarily for official use.
+However, they can be helpful if you want to implement your own catalog.
+So use them at your own risk.
