+++ pkg/cldrplural/ast/scanner.go
