+++ pkg/cldrplural/ast/scanner_test.go
