+++ pkg/cldrplural/generator/generate.sh
