+++ pkg/poplural/ast/scanner.go
