+++ pkg/poplural/generator/.gitignore
