+++ pkg/poplural/generator/compiler.go
@@ -4,7 +4,7 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/vorlif/spreak/internal/poplural/ast"
+	"github.com/vorlif/spreak/pkg/poplural/ast"
 )
 
 func compileForms(forms *ast.Forms) string {
@@ -16,7 +16,7 @@ func compileForms(forms *ast.Forms) string {
 
 	if forms.NPlurals == 2 {
 		b.WriteString("if ")
-		compileTwoPluralNode(forms.Root, &b)
+		compileBinaryPluralNode(forms.Root, &b)
 		b.WriteString("{\nreturn 1\n}\n")
 		b.WriteString("return 0")
 	} else {
@@ -26,7 +26,7 @@ func compileForms(forms *ast.Forms) string {
 	return b.String()
 }
 
-func compileTwoPluralNode(node ast.Node, b *strings.Builder) {
+func compileBinaryPluralNode(node ast.Node, b *strings.Builder) {
 	switch v := node.(type) {
 	case *ast.ValueExpr:
 		b.WriteString(strconv.FormatInt(v.Value, 10))
@@ -35,43 +35,43 @@ func compileTwoPluralNode(node ast.Node, b *strings.Builder) {
 	case *ast.BinaryExpr:
 		switch v.Type() {
 		case ast.LogicalAnd:
-			compileTwoPluralNode(v.X, b)
+			compileBinaryPluralNode(v.X, b)
 			b.WriteString(" && ")
-			compileTwoPluralNode(v.Y, b)
+			compileBinaryPluralNode(v.Y, b)
 		case ast.LogicalOr:
 			b.WriteString("(")
-			compileTwoPluralNode(v.X, b)
+			compileBinaryPluralNode(v.X, b)
 			b.WriteString(" || ")
-			compileTwoPluralNode(v.Y, b)
+			compileBinaryPluralNode(v.Y, b)
 			b.WriteString(")")
 		case ast.Equal:
-			compileTwoPluralNode(v.X, b)
+			compileBinaryPluralNode(v.X, b)
 			b.WriteString(" == ")
-			compileTwoPluralNode(v.Y, b)
+			compileBinaryPluralNode(v.Y, b)
 		case ast.NotEqual:
-			compileTwoPluralNode(v.X, b)
+			compileBinaryPluralNode(v.X, b)
 			b.WriteString(" != ")
-			compileTwoPluralNode(v.Y, b)
+			compileBinaryPluralNode(v.Y, b)
 		case ast.Greater:
-			compileTwoPluralNode(v.X, b)
+			compileBinaryPluralNode(v.X, b)
 			b.WriteString(" > ")
-			compileTwoPluralNode(v.Y, b)
+			compileBinaryPluralNode(v.Y, b)
 		case ast.GreaterOrEqual:
-			compileTwoPluralNode(v.X, b)
+			compileBinaryPluralNode(v.X, b)
 			b.WriteString(" >= ")
-			compileTwoPluralNode(v.Y, b)
+			compileBinaryPluralNode(v.Y, b)
 		case ast.Less:
-			compileTwoPluralNode(v.X, b)
+			compileBinaryPluralNode(v.X, b)
 			b.WriteString(" < ")
-			compileTwoPluralNode(v.Y, b)
+			compileBinaryPluralNode(v.Y, b)
 		case ast.LessOrEqual:
-			compileTwoPluralNode(v.X, b)
+			compileBinaryPluralNode(v.X, b)
 			b.WriteString(" <= ")
-			compileTwoPluralNode(v.Y, b)
+			compileBinaryPluralNode(v.Y, b)
 		case ast.Reminder:
-			compileTwoPluralNode(v.X, b)
+			compileBinaryPluralNode(v.X, b)
 			b.WriteString("%")
-			compileTwoPluralNode(v.Y, b)
+			compileBinaryPluralNode(v.Y, b)
 		}
 	}
 }
@@ -81,7 +81,7 @@ func compileNode(node ast.Node, b *strings.Builder) {
 	switch v := node.(type) {
 	case *ast.QuestionMarkExpr:
 		b.WriteString("if ")
-		compileTwoPluralNode(v.Cond, b)
+		compileBinaryPluralNode(v.Cond, b)
 		if vn, ok := v.T.(*ast.ValueExpr); ok {
 			b.WriteString("{\n")
 			b.WriteString("return ")
@@ -98,6 +98,6 @@ func compileNode(node ast.Node, b *strings.Builder) {
 			compileNode(v.F, b)
 		}
 	default:
-		compileTwoPluralNode(v, b)
+		compileBinaryPluralNode(v, b)
 	}
 }
